fix(feed): share platform normalization and fix tab callback typing
